+++ vgui2/vgui_controls/ImagePanel.cpp
@@ -85,7 +85,7 @@ void ImagePanel::SetImage(const char *imageName)
 	if ( imageName && m_pszImageName && V_stricmp( imageName, m_pszImageName ) == 0 )
 		return;
 
-	int len = Q_strlen(imageName) + 1;
+	intp len = Q_strlen(imageName) + 1;
 	delete [] m_pszImageName;
 	m_pszImageName = new char[ len ];
 	Q_strncpy(m_pszImageName, imageName, len );
@@ -305,7 +305,7 @@ void ImagePanel::ApplySettings(KeyValues *inResourceData)
 	if (*pszFillColor)
 	{
 		int r = 0, g = 0, b = 0, a = 255;
-		int len = Q_strlen(pszFillColor) + 1;
+		intp len = Q_strlen(pszFillColor) + 1;
 		m_pszFillColorName = new char[ len ];
 		Q_strncpy( m_pszFillColorName, pszFillColor, len );
 
@@ -329,7 +329,7 @@ void ImagePanel::ApplySettings(KeyValues *inResourceData)
 	if (*pszDrawColor)
 	{
 		int r = 255, g = 255, b = 255, a = 255;
-		int len = Q_strlen(pszDrawColor) + 1;
+		intp len = Q_strlen(pszDrawColor) + 1;
 		m_pszDrawColorName = new char[ len ];
 		Q_strncpy( m_pszDrawColorName, pszDrawColor, len );
